Added $ok property for identifying success responses

+++ server.php
@@ -4,6 +4,7 @@ require_once __DIR__ .'/bootstrap/app.php';
 
 
 use Amp\ByteStream\ResourceOutputStream;
+use Amp\Delayed;
 use Amp\Http\Server\RequestHandler\CallableRequestHandler;
 use Amp\Http\Server\Response;
 use Amp\Http\Server\Router;
@@ -38,10 +39,15 @@ Amp\Loop::run(function () {
 
         $fn = fn($id) => call(function () use ($api, $id) {
             $message = yield $api->sendMessage(-1001187469156, var_export('Test', true));
-//            yield new Delayed(1000);
-            $edited = yield $message->edit('Hi');
-//            yield new Delayed(1000);
-            return $edited->delete();
+            if ($message->isOk()) {
+                yield new Delayed(1000);
+                $edited = yield $message->edit('Hi');
+                if ($edited->isOk()) {
+                    yield new Delayed(1000);
+                    return $edited->delete();
+                }
+            }
+
         });
 
         foreach (range(1, 5) as $i)
